Style: #6 - 메인페이지 개발을 위해 지도 컴포넌트 분리
